Huge code refactoring to remove urwid functionality
This commit removes all functionality designed to present the user with
a menu with which to select a URL. The aim is to turn this into a module
that does nothing but extract URLs and output them. I believe this is
all it should have done from the beginning, and the urwid-based
functionality should have gone into a separate module.